eaglercraft chrome extension
"To access additional features, you can use this version of Eaglercraft as a Chrome extension. Follow these steps to get started:

Create a new folder on your computer.
Copy all the necessary files into the folder, including the Eaglercraft client you are currently using.
Rename the Eaglercraft client file to 'eaglercraft'.
Open your Chrome browser and go to the settings.
Enable the developer mode for Chrome extensions.
Select 'Load Unpacked' and choose the folder where you placed all the files, including the 'eaglercraft' launcher.
Eaglercraft will now be added as a Chrome extension.
Please note that this version is super new and the extension is still not working yet.